requirements.txt
streamlit
tensorflow
matplotlib
pandas
numpy
seaborn
pillow
time
